refactored names and directories
